add upcoming only checkbox and split views/tests
