+++ templates/event/event_list.html
@@ -18,6 +18,11 @@
                 <label class="fr-label" for="{{ form.scale.id_for_label }}">{{ form.scale.label }}</label>
                 {{ form.scale }}
               </div>
+              <div class="fr-checkbox-group">
+                {{ form.upcoming.errors }}
+                {{ form.upcoming }}
+                <label class="fr-label" for="{{ form.upcoming.id_for_label }}">{{ form.upcoming.label }}</label>
+              </div>
             </form>
           </div>
           <div class="fr-col-9">
@@ -30,7 +35,7 @@
                         <h3 class="fr-card__title">
                           <a href="{% url 'event_detail' event.pk %}">{{ event.subject }}</a>
                         </h3>
-                        <p class="fr-card__desc">{{ event.description }}</p>
+                        <p class="fr-card__desc"></p>
                         <div class="fr-card__start">
                           <ul class="fr-badges-group">
                             <li>
