+++ templates/event/organizer/dashboard.html
@@ -20,14 +20,14 @@
             <a href="{% url 'event_organizer_event_create' %}" class="fr-btn" target="_self">Créer une nouvelle concertation</a>
           </div>
           {% for event in event_list %}
-            <div class="fr-col-md-6 fr-col">
+            <div class="fr-col-md-4 fr-col">
               <div class="fr-card fr-enlarge-link">
                 <div class="fr-card__body">
                   <div class="fr-card__content">
                     <h3 class="fr-card__title">
                       <a href="{% url 'event_organizer_event_detail' event.pk %}">{{ event.subject }}</a>
                     </h3>
-                    <p class="fr-card__desc">{{ event.description }}</p>
+                    <p class="fr-card__desc"></p>
                     <div class="fr-card__start">
                       <ul class="fr-badges-group">
                         <li>
